fix up app js top code and move my Components to components
